requirements.txt
torch==2.0.1
numpy==1.25.0
tqdm==4.65.0
transformers==4.33.2
lightning==2.0.9
pillow==9.5.0
easydict==1.10